// src/enums/ERoles.ts
export enum ERoles {
    // WEREWOLVES
    WEREWOLF = "WEREWOLF",
    INFECT_WEREWOLF = "INFECT_WEREWOLF",

    // VILLAGERS
    SIMPLE_VILLAGER = "SIMPLE_VILLAGER",
    WITCH = "WITCH",
    LITTLE_GIRL = "LITTLE_GIRL",
    GUARD = "GUARD",
    CUPID = "CUPID",
    SEER = "SEER",
    HUNTER = "HUNTER",

    // SOLOS
    WHITE_WEREWOLF = "WHITE_WEREWOLF"
}